deep recursion test + improvements (#45)
* MrBot wrote some tests

* added a deep call stack test

* Fix all C++ compiler warnings and modernize code

- criticalsection.h: Replace internal CRITICAL_SECTION struct access with
  explicit thread ID tracking for IsLockedByCurrentThread()
- utility.cpp: Remove deprecated 'register' storage class keyword
- Allocs.cpp: Replace deprecated std::wstring_convert/codecvt with
  WideCharToMultiByte() Windows API
- basics.cpp, dynamic_app.cpp: Update deprecated INSTANTIATE_TEST_CASE_P
  to INSTANTIATE_TEST_SUITE_P (gtest modernization)
- testsuite.cpp: Fix signed/unsigned mismatch by using UINT/ULONG types
- Remove redundant VLD_FORCE_ENABLE macro definitions (already defined
  via CMake for non-Debug builds)
- Add deep_callstack_test for testing long stack trace handling

Both ARM64 and x64 builds now compile with zero warnings.

* Add VLD_WARNINGS_AS_ERRORS CMake option

Enable treating compiler warnings as errors by default (/WX flag).
This ensures code quality by preventing warnings from accumulating.

Can be disabled with: cmake .. -DVLD_WARNINGS_AS_ERRORS=OFF

* Upgrade to warning level 4 (/W4)

- CMakeLists.txt: Change /W3 to /W4 for stricter warnings
- crtmfcpatch.h: Suppress C4201 for intentional nameless struct/union
- callstack.cpp: Suppress C4127 for third-party cppformat library header
- utility.cpp: Wrap VLDVirtualProtect/Restore in #if !defined(_M_ARM64)
  since they're only used for x86/x64 trampoline detection
- utility.cpp, vld.cpp: Fix unused variable warnings with (void) cast

All warnings fixed consistently using (void)var pattern for debug-only
variables. Both ARM64 and x64 build clean with /W4 /WX.

* Upgrade cppformat to {fmt} 12.1.1

- Replace ancient cppformat 1.1.0 (March 2015) with {fmt} 12.1.1
- Update callstack.cpp to use modern fmt::format_to_n() API
- Configure fmt library for static runtime to match VLD
- Suppress warnings-as-errors for third-party fmt code
- Eliminates deprecated stdext::checked_array_iterator warnings

* Fix wcstombs_s truncation handling in Print()

STRUNCATE return value from wcstombs_s with _TRUNCATE flag is not an
error - it indicates successful truncation. Only assert on actual
conversion failures, not on truncated output.

This fixes Debug test failures with deep call stacks that exceed
the MAXMESSAGELENGTH buffer size.

* Fix LNK4281 warning: use base address >= 4GB for x64

Set DLL base address to 0x100000000 (4GB) for both x64 and ARM64
platforms for optimal ASLR (Address Space Layout Randomization).
Keep 0x03200000 only for x86 (32-bit) which cannot address above 4GB.

* Add linker warnings as errors (/WX)

* Add detailed documentation to deep_callstack_test explaining buffer size calculation

* Remove unnecessary C4127 warning suppression for fmt

* Restore C4127 warning suppression for fmt header

* Fix ARM64 build to use ARM64-hosted compiler

* Fix vcvarsall argument: use arm64_arm64 for ARM64-hosted ARM64 target

* Revert to arm64 vcvarsarg - ARM64-hosted tools not available in VS image

+++ src/tests/suite/testsuite.cpp
@@ -100,9 +100,9 @@ enum action_e {
 #endif
 #endif
 
-static int MAXALLOC = 0;                     // Maximum number of allocations of each type to perform, per thread
-static int NUMTHREADS = 0;                   // Number of threads to run simultaneously
-static int MAXBLOCKS = 0;                    // Total maximum number of allocations, per thread
+static UINT MAXALLOC = 0;                     // Maximum number of allocations of each type to perform, per thread
+static UINT NUMTHREADS = 0;                   // Number of threads to run simultaneously
+static ULONG MAXBLOCKS = 0;                    // Total maximum number of allocations, per thread
 
 static const int MAXDEPTH = 32;              // Maximum depth of the allocation call stack
 static const int MAXSIZE = 64;               // Maximum block size to allocate
